:truck: Refactor : 카프카, 레디스 설정파일 위치 변경
